+++ e2e/global-teardown.ts
@@ -2,6 +2,8 @@
 import { execSync } from "child_process";
 
 export default async function globalTeardown() {
+  if (process.env.CI) return;
+
   execSync("docker-compose -f e2e/docker-compose.test.yml down -v", {
     stdio: "inherit",
   });
